Caches downloads first and then loads them into database.
Initially I was trying to download a chunk of xml from SRA and then load
that chunk into the database. However, it appears there is a timeout on
the webenv; if the loading takes too long I begin to get IncompleteRead
errors. To try to fix this I changed things up to download and cache all
of the chunks and then load them into the database. I still have the
occassional server error, but the user just needs to re-run the query
and it should start where it left off.